Stable Beta Version 2. Flask Authentication Implemented

+++ todolist/__init__.py
@@ -1,6 +1,8 @@
-from flask import Flask, render_template, request,g
+from flask import Flask, render_template, request, g
+from flask_login import current_user, LoginManager
 import datetime
 import os
+from .db import User
 
 today_date = datetime.datetime.today().strftime ('%Y-%m-%d')
 
@@ -32,33 +34,75 @@ def create_app(test_config=None):
   
   from . import db 
   db.init_app(app)
+
+  from .auth import auth as auth_blueprint
+  app.register_blueprint(auth_blueprint)
+
+  from .main import main as main_blueprint
+  app.register_blueprint(main_blueprint)
+
+  login_manager = LoginManager()
+  login_manager.init_app(app)
+
+  @login_manager.user_loader
+  def load_user(user_id):
+      # Implement a function to load the user from your database
+      # Example: return User.query.get(int(user_id))
+      conn = db.get_db()
+      cursor = conn.cursor()
+      cursor.execute(f"select * from user where id = %s", (user_id, ))
+      user = cursor.fetchone()
+      return user
+
+
+  @app.before_request
+  def before_request():
+    g.user_id = current_user.id if current_user.is_authenticated else None
   
-  @app.route("/")
-  def index():
+  @login_manager.user_loader
+  def load_user(user_id):
     conn = db.get_db()
     cursor = conn.cursor()
-    cursor.execute(f"insert into overdue_tasks (task, points) select t.task, t.points from task t where t.date_of_task < %s and t.status=%s", (today_date, 2))
-    cursor.execute(f"delete from task t where t.date_of_task < %s", (today_date, ))
+    cursor.execute(f"SELECT * FROM app_user WHERE id = %s", (user_id,))
+    user_data = cursor.fetchone()
+
+    if user_data:
+        user = User(*user_data)
+        return user
+    return None
+
+
+  #user_id = g.user_id
+
+  
+  @app.route("/index")
+  def index():
+    g.user_id = current_user.id if current_user.is_authenticated else None
+    user_id = g.user_id
+    conn = db.get_db()
+    cursor = conn.cursor()
+    cursor.execute(f"insert into overdue_tasks (task, points) select t.task, t.points from task t where t.date_of_task < %s and t.status_id=%s and t.user_id = %s", (today_date, 2, user_id))
+    cursor.execute(f"delete from task t where t.date_of_task < %s and t.user_id = %s", (today_date, user_id))
     conn.commit()
     oby = request.args.get("order_by", "date_of_task")  
     order = request.args.get("order", "asc")
     if oby == "date_of_task":
       if order == "asc":
-           cursor.execute(f"select t.id, t.task, t.date_of_task, t.day, u.urgency, ts.status from task t, urgency u, task_status ts where t.urgency=u.id and t.status=ts.id and t.status = 2 order by t.date_of_task")
+           cursor.execute(f"select t.id, t.task, t.date_of_task, t.day, u.urgency, ts.status from task t, urgency u, task_status ts where t.urgency_id=u.id and t.status_id=ts.id and t.status_id = 2 and t.user_id = %s order by t.date_of_task", (user_id,))
       else:
-           cursor.execute(f"select t.id, t.task, t.date_of_task, t.day, u.urgency, ts.status from task t, urgency u, task_status ts where t.urgency=u.id and t.status=ts.id and t.status = 2 order by t.date_of_task desc")
+           cursor.execute(f"select t.id, t.task, t.date_of_task, t.day, u.urgency, ts.status from task t, urgency u, task_status ts where t.urgency_id=u.id and t.status_id=ts.id and t.status_id = 2 and t.user_id = %s order by t.date_of_task desc", (user_id,))
     if oby == "urgency":
       if order == "asc":
-           cursor.execute(f"select t.id, t.task, t.date_of_task, t.day, u.urgency, ts.status from task t, urgency u, task_status ts where t.urgency=u.id and t.status=ts.id and t.status = 2 order by t.urgency")
+           cursor.execute(f"select t.id, t.task, t.date_of_task, t.day, u.urgency, ts.status from task t, urgency u, task_status ts where t.urgency_id=u.id and t.status_id=ts.id and t.status_id = 2 and t.user_id = %s order by t.urgency_id", (user_id,))
       else:
-           cursor.execute(f"select t.id, t.task, t.date_of_task, t.day, u.urgency, ts.status from task t, urgency u, task_status ts where t.urgency=u.id and t.status=ts.id and t.status = 2 order by t.urgency desc")
+           cursor.execute(f"select t.id, t.task, t.date_of_task, t.day, u.urgency, ts.status from task t, urgency u, task_status ts where t.urgency_id=u.id and t.status_id=ts.id and t.status_id = 2 and t.user_id = %s order by t.urgency_id desc", (user_id,))
     if oby == "status":
       if order == "asc":
-           cursor.execute(f"select t.id, t.task, t.date_of_task, t.day, u.urgency, ts.status from task t, urgency u, task_status ts where t.urgency=u.id and t.status=ts.id and t.status = 2 order by t.date_of_task")
+           cursor.execute(f"select t.id, t.task, t.date_of_task, t.day, u.urgency, ts.status from task t, urgency u, task_status ts where t.urgency_id=u.id and t.status_id=ts.id and t.status_id = 2 and t.user_id = %s order by t.date_of_task", (user_id,))
       else:
-           cursor.execute(f"select t.id, t.task, t.date_of_task, t.day, u.urgency, ts.status from task t, urgency u, task_status ts where t.urgency=u.id and t.status=ts.id and t.status = 1 order by t.date_of_task")
+           cursor.execute(f"select t.id, t.task, t.date_of_task, t.day, u.urgency, ts.status from task t, urgency u, task_status ts where t.urgency_id=u.id and t.status_id=ts.id and t.status_id = 1 and t.user_id = %s order by t.date_of_task", (user_id,))
     tasks = cursor.fetchall()
-    cursor.execute(f"select o.id, o.task from overdue_tasks o")
+    cursor.execute(f"select o.id, o.task from overdue_tasks o where o.user_id =%s", (user_id,))
     overdue = cursor.fetchall()
     
     return render_template('index.html', tasks=tasks, overdue=overdue, order="desc" if order=="asc" else "asc")
@@ -69,14 +113,11 @@ def create_app(test_config=None):
     endweek_date = (datetime.datetime.today() + datetime.timedelta(days=7)).strftime ('%Y-%m-%d')
     conn = db.get_db()
     cursor = conn.cursor()
-    cursor.execute("select t.id, t.task, t.date_of_task, t.day, u.urgency, ts.status from task t, urgency u, task_status ts where t.urgency=u.id and t.status=ts.id and t.date_of_task between (%s) and (%s) order by t.date_of_task", (today_date, endweek_date))
+    cursor.execute("select t.id, t.task, t.date_of_task, t.day, u.urgency, ts.status from task t, urgency u, task_status ts where t.urgency_id=u.id and t.status_id=ts.id and t.date_of_task between (%s) and (%s) order by t.date_of_task", (today_date, endweek_date))
     tasks = cursor.fetchall()
     return render_template('index.html', tasks=tasks)
     
-  #@app.route('/task/<int:id>/delete', methods=['GET', 'POST'])
-  #def delete_task(id):
-  #  cursor.execute("delete from task where id = %d", (id))
-  #  return redirect(url_for('index'))
+ 
 
   return app
